% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{available_models}
\alias{available_models}
\title{List available Stan models}
\usage{
available_models(score = "")
}
\arguments{
\item{score}{(optional) score to subset the output}
}
\value{
Dataframe with columns Score and Model
}
\description{
NB: does not include uniform or historical forecast
}
