% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{custom_metrics}
\alias{custom_metrics}
\alias{compute_accuracy}
\alias{compute_quantile_error}
\title{Compute accuracy and quantile error}
\usage{
compute_accuracy(y, pred, ct)

compute_quantile_error(y, pred, p)
}
\arguments{
\item{y}{vector of outcomes}

\item{pred}{list of same size of y containing vector of samples}

\item{ct}{cut-off}

\item{p}{probability value}
}
\value{
Vector of accuracy values
}
\description{
\itemize{
\item "Accuracy" is defined as the probability that absolute error less than specified value
\item "Quantile error" is defined as the quantile p of the predictive error distribution
}
}
\details{
The functions are designed to take a vector of outcomes and a list of samples, so they can be used with a dataframe.
An alternative would be to use \code{purrr::map2} in \code{mutate} directly, and define functions that take a scalar for the outcome
and a vector for the samples:
e.g. \code{map2(Score, Samples, ~mean(abs(.x - .y) < acc_thr)) \%>\% unlist()}
}
