% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{detail_POSCORAD}
\alias{detail_POSCORAD}
\title{List and details available severity score and items in the Derexyl and PFDC dataset}
\usage{
detail_POSCORAD(filter_output = NULL)
}
\arguments{
\item{filter_output}{Optional character vector to subset the output by the "Name" column.
It can also take the value "Components" to select A (extent), B and C;
"Scores" to select SCORAD and oSCORAD;
"Items" to select all severity items;
"Intensity signs" to select all intensity signs;
and "Subjective symptoms" to select all subjective symptoms.
If \code{length(filter_output) > 1}, it acts as a OR.}
}
\value{
Dataframe with columns:
\itemize{
\item Name: short name of the severity item/score
\item Label: full name of the severity item/score in the dataset
\item Maximum: maximum value that the severity item/score can take
\item Resolution: resolution of the score
}
}
\description{
List and details available severity score and items in the Derexyl and PFDC dataset
}
\examples{
detail_POSCORAD()
detail_POSCORAD("Items")
detail_POSCORAD(c("B", "C", "oSCORAD", "SCORAD"))
}
