% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EczemaPredPOSCORAD-package.R
\docType{package}
\name{EczemaPredPOSCORAD-package}
\alias{EczemaPredPOSCORAD}
\alias{EczemaPredPOSCORAD-package}
\title{EczemaPredPOSCORAD: Predicting PO-SCORAD Using EczemaPred}
\description{
Analysis package investigating the prediction of PO-SCORAD severity score for eczema.
}
\author{
\strong{Maintainer}: Guillem Hurault \email{guillem.hurault@hotmail.fr} (\href{https://orcid.org/0000-0002-1052-3564}{ORCID})

}
\keyword{internal}
