% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{estimate_performance}
\alias{estimate_performance}
\title{Estimate performance (learning curve) with a meta-model}
\usage{
estimate_performance(metric_name, df, fc_it, adjust_horizon = TRUE)
}
\arguments{
\item{metric_name}{Name of the metric to estimate the learning curve for}

\item{df}{Dataframe of validation results}

\item{fc_it}{Forward chaining iteration characteristics}

\item{adjust_horizon}{Whether to adjust for prediction horizon in the model}
}
\value{
Dataframe
}
\description{
The meta-model is a Generative Additive Mixed effect model
}
