% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FakeData}
\alias{FakeData}
\title{Fake PO-SCORAD dataset}
\format{
A named list with element \code{Data} and \code{Parameters} containing dataframes.
The dataframe \code{Data} contains the PO-SCORAD time-series.
It has the following columns:
\describe{
\item{Patient}{Patient ID}
\item{Day}{Day since first measurement (day=1 is the first measurement)}
\item{Extent}{}
\item{Itching VAS}{}
\item{Sleep disturbance VAS}{}
\item{Dryness}{}
\item{Redness}{}
\item{Swelling}{}
\item{Scabs/Oozing}{}
\item{Traces of scratching}{}
\item{Thickening}{}
\item{Intensity}{Dryness + Redness + Swelling + Scabs/Oozing + Traces of scratching + Thickening}
\item{Subjective symptoms}{Itching VAS + Sleep disturbance VAS}
\item{oSCORAD}{0.2 * Extent + 3.5 * Intensity}
\item{SCORAD}{0.2 * Extent + 3.5 * Intensity + Subjective symptoms}
}
The dataframe \code{Parameters} contains the parameters' values of the model used to generate the data.
It has the following columns:
\describe{
\item{Model}{Name of the model}
\item{Item}{Name of the item}
\item{Draw}{Draw used to generate the data}
\item{Index}{Index of the parameter}
\item{Value}{Parameter value}
\item{Parameter}{Name of the parameter}
}
}
\usage{
FakeData
}
\description{
The items are generated independently of each other.
\itemize{
\item Extent is simulated with a "BinMC" model
\item intensity signs are generated with an "OrderedRW" model
\item subjective symptoms are generated with an "BinRW" model
}
}
\keyword{datasets}
