% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_management.R
\name{get_results_files}
\alias{get_results_files}
\title{Dictionary of results files}
\usage{
get_results_files(outcome, model, dataset = NULL, val_horizon = NULL)
}
\arguments{
\item{outcome}{item/score to predict}

\item{model}{model name}

\item{dataset}{dataset name}

\item{val_horizon}{validation horizon}
}
\value{
Named list:
\itemize{
\item Prior: Stanfit object of prior predictive distribution
\item PriorPar: Dataframe of posterior summary statistics of parameters
\item FakeFit: Stanfit object of fake data fit
In addition, if \code{dataset} is not NULL:
\item Fit: Stanfit object of the fit
\item FitPar: Dataframe of posterior summary statistics of parameters
In addition, if \code{val_horizon} is not NULL:
\item ValDir: Temporary directory storing individual iteration results
\item Val: Full validation results
}
}
\description{
Dictionary of results files
}
