% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_horizon_change}
\alias{plot_horizon_change}
\title{Plot performance change as prediction horizon is increased}
\usage{
plot_horizon_change(perf)
}
\arguments{
\item{perf}{Dataframe of performance estimates}
}
\value{
Ggplot
}
\description{
Plot performance change as prediction horizon is increased
}
