% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{plot_perf_vs_score}
\alias{plot_perf_vs_score}
\title{Plot performance (eg lpd_diff) vs score}
\usage{
plot_perf_vs_score(perf, metric = "lpd_diff", discrete = FALSE, max_score = NA)
}
\arguments{
\item{perf}{Dataframe of performance estimates}

\item{metric}{Name of the metric to plot}

\item{discrete}{Whether to plot as if y was a discrete outcome or a continuous one (loess is then applied)}

\item{max_score}{Maximum value that the score can take}
}
\value{
Ggplot
}
\description{
Plot performance (eg lpd_diff) vs score
}
